refactor(styles): move styles-v2 into styles/, keep v1 globals.css
- git mv src/styles-v2/{tokens.css,global.css,index.css,accent.ts,
  components/,pages/} -> src/styles/
- rmdir src/styles-v2
- sed: replace 'styles-v2' with 'styles' across 14 src/ files (imports +
  css internal @import + comment paths)

src/styles/globals.css (v1) is intentionally retained — its utility classes
(.btn-*, .form-*, .stat-card, .container, .table, .badge, .tabs,
.pagination, .divider, .empty-state, .toast, .search-bar) are still used
by v2 Landing (HeroSection / CtaSection / Stat) and 13 v1 retained pages
(Signup, Payment, SellerApply, NotFound, OAuthCallback, SignupComplete,
SocialProfileSetup, WalletCharge*, admin/*, seller/*).

main.tsx CSS load order preserved: styles/index.css (v2 barrel) before
styles/globals.css (v1 utilities) — same as before the move.

Plan §10.2.3 commit 8's "git rm globals.css" deferred until v1 retained
pages get a redesign track of their own.
