+++ src/pages/Cart/components/CartHeader.tsx
@@ -2,7 +2,7 @@
  * 페이지 상단 타이틀 + 부제.
  *
  * 프로토타입 Cart.jsx:19-22 의 인라인 스타일을 BEM 클래스로 변환.
- * 시각: h1 26px/700 text, 부제 14px text-3 (CSS는 styles-v2/pages/cart.css).
+ * 시각: h1 26px/700 text, 부제 14px text-3 (CSS는 styles/pages/cart.css).
  */
 
 export interface CartHeaderProps {
