+++ src/styles/components/accent-media-box.css
