+++ src/styles/components/empty-state.css
