+++ src/styles/components/ide-chrome.css
