+++ src/styles/components/section-head.css
