+++ src/styles/components/status-chip.css
