+++ src/styles/components/term-dot.css
