+++ src/styles/pages/event-detail.css
