+++ src/styles/pages/mypage-tickets.css
