+++ src/styles/pages/payment-callback.css
